requirements.txt
mysql-connector-python
Flask==1.1.2
flask-restful==0.3.9
pyjwt
itsdangerous==1.1.0
Jinja2==2.11.3
markupsafe==2.0.1
werkzeug==2.0.3